chore: #657 vega tx dialog used for order cancellation and order submission
