+++ apps/trading/pages/_app.page.tsx
@@ -93,7 +93,7 @@ function VegaTradingApp(props: AppProps) {
           <link rel="stylesheet" href="https://static.vega.xyz/fonts.css" />
           {['1', 'true'].includes(process.env['NX_USE_ENV_OVERRIDES'] || '') ? (
             /* eslint-disable-next-line @next/next/no-sync-scripts */
-            <script src="./env-config.js" type="text/javascript" />
+            <script src="./env-config.js" type="text/javascript"></script>
           ) : null}
         </Head>
         <AppBody {...props} />
