+++ libs/deal-ticket/src/components/deal-ticket-manager.tsx
@@ -1,8 +1,10 @@
 import type { ReactNode } from 'react';
-import { useEffect, useState } from 'react';
-import { Dialog, Intent } from '@vegaprotocol/ui-toolkit';
-import { OrderStatus } from '@vegaprotocol/types';
-import { VegaOrderTransactionDialog, VegaTxStatus } from '@vegaprotocol/wallet';
+import { useState } from 'react';
+import {
+  VegaOrderTransactionType,
+  VegaTransactionDialog,
+  VegaTxStatus,
+} from '@vegaprotocol/wallet';
 import { DealTicket } from './deal-ticket';
 import { useOrderSubmit } from '../hooks/use-order-submit';
 import type { DealTicketQuery_market } from './__generated__/DealTicketQuery';
@@ -19,42 +21,6 @@ export const DealTicketManager = ({
   const [orderDialogOpen, setOrderDialogOpen] = useState(false);
   const { submit, transaction, finalizedOrder, reset } = useOrderSubmit(market);
 
-  const getDialogIntent = (status: VegaTxStatus) => {
-    if (finalizedOrder) {
-      if (
-        finalizedOrder.status === OrderStatus.Active ||
-        finalizedOrder.status === OrderStatus.Filled ||
-        finalizedOrder.status === OrderStatus.PartiallyFilled
-      ) {
-        return Intent.Success;
-      }
-
-      if (finalizedOrder.status === OrderStatus.Parked) {
-        return Intent.Warning;
-      }
-
-      return Intent.Danger;
-    }
-
-    if (status === VegaTxStatus.Requested) {
-      return Intent.Warning;
-    }
-
-    if (status === VegaTxStatus.Error) {
-      return Intent.Danger;
-    }
-
-    return Intent.None;
-  };
-
-  useEffect(() => {
-    if (transaction.status !== VegaTxStatus.Default || finalizedOrder) {
-      setOrderDialogOpen(true);
-    } else {
-      setOrderDialogOpen(false);
-    }
-  }, [finalizedOrder, transaction.status]);
-
   return (
     <>
       {children || (
@@ -69,23 +35,14 @@ export const DealTicketManager = ({
           }
         />
       )}
-      <Dialog
-        open={orderDialogOpen}
-        onChange={(isOpen) => {
-          setOrderDialogOpen(isOpen);
-
-          // If closing reset
-          if (!isOpen) {
-            reset();
-          }
-        }}
-        intent={getDialogIntent(transaction.status)}
-      >
-        <VegaOrderTransactionDialog
-          transaction={transaction}
-          finalizedOrder={finalizedOrder}
-        />
-      </Dialog>
+      <VegaTransactionDialog
+        orderDialogOpen={orderDialogOpen}
+        setOrderDialogOpen={setOrderDialogOpen}
+        finalizedOrder={finalizedOrder}
+        transaction={transaction}
+        reset={reset}
+        type={VegaOrderTransactionType.SUBMIT}
+      />
     </>
   );
 };
