--- libs/orders/src/lib/components/cancel-order-dialog/index.ts
@@ -1,1 +1,0 @@
-export * from './cancel-dialog';
