+++ libs/orders/src/lib/components/index.ts
@@ -1,5 +1,4 @@
 export * from './__generated__';
-export * from './cancel-order-dialog';
 export * from './mocks';
 export * from './order-data-provider';
 export * from './order-list';
