+++ libs/wallet/src/index.ts
@@ -7,5 +7,6 @@ export * from './use-vega-transaction';
 export * from './use-eager-connect';
 export * from './manage-dialog';
 export * from './vega-order-transaction-dialog';
+export * from './vega-transaction-dialog';
 export * from './provider';
 export * from './connect-dialog';
