+++ libs/wallet/src/vega-transaction-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './vega-transaction-dialog';
